Added previous version (Mridul's Version) of the scripts in the project for reference.
Fixed all the errors in previous scripts and renamed them for clear separation from new scripts.

Current version of Scripts are in the original state which will need to be updated later

This commit includes some project files and some changes corresponding to the use of Rider (Code IDE)

The script Drive.cs was also updated and refactored.

The script folder is reorganized as well.

The scene has been moved inside a folder.